Provide `tag` flag only if variable is set
We were trying to execute a command when version-tag input was not set.
I caused empty `--tag` to be provided which errored with tenderly.
Here we moved the push command to a script and check for required
environment variables for ETH_NETWORK_ID and TENDERLY_PROJECT_SLUG.
Tags are optional and only if tag value is present it's appended to the
push command.